refactor(android): group request-close implementation
