+++ android/src/test/java/com/swmansion/reactnativebottomsheet/requestclose/PortalRequestCloseRoutingContextTest.kt
@@ -1,4 +1,4 @@
-package com.swmansion.reactnativebottomsheet
+package com.swmansion.reactnativebottomsheet.requestclose
 
 import android.app.Activity
 import android.os.Looper
