+++ android/src/test/java/com/swmansion/reactnativebottomsheet/requestclose/RequestCloseInputTest.kt
@@ -1,4 +1,4 @@
-package com.swmansion.reactnativebottomsheet
+package com.swmansion.reactnativebottomsheet.requestclose
 
 import android.view.KeyEvent
 import androidx.activity.BackEventCompat
